Messing with my flag handling

+++ ft_printf.c
@@ -6,7 +6,7 @@
 /*   By: maljean <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/05/29 23:48:08 by maljean           #+#    #+#             */
-/*   Updated: 2018/06/01 18:21:07 by maljean          ###   ########.fr       */
+/*   Updated: 2018/06/01 18:21:49 by maljean          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -479,7 +479,7 @@ int	ft_strcmp(const char *s1, const char *s2)
 char	*ft_mod_cast(va_list args, t_ops *ops, int base)
 {
 	if (!ops->mod)
-		return (ft_itoabase_umax(va_arg(args, ssize_t), base, ops));
+		return (ft_itoabase_umax(va_arg(args, int), base, ops));
 	else if (ft_strcmp(ops->mod, "l") && ops->conv == 'd')
 		return (ft_itoabase_umax(va_arg(args, long), base, ops));
 	else if (ft_strcmp(ops->mod, "j") || ft_strcmp(ops->mod, "z"))
